feat: verify token for openbanking
